(PC-17810) refactor(bookOffer): move DuoChoice, HourChoice and ChoiceBloc to components folder
