+++ src/features/bookOffer/components/DuoChoice.tsx
@@ -1,10 +1,9 @@
 import React, { FunctionComponent } from 'react'
 import styled from 'styled-components/native'
 
+import { ChoiceBloc, getTextColor } from 'features/bookOffer/components/ChoiceBloc'
 import { IconInterface } from 'ui/svg/icons/types'
 import { getSpacing, Typo } from 'ui/theme'
-
-import { ChoiceBloc, getTextColor } from './ChoiceBloc'
 
 interface Props {
   title: string
